restructure the project folders
