+++ common_components/SelectDrop.js
@@ -1,5 +1,5 @@
-import { StyleSheet, Text, View } from "react-native";
-import React, { useState } from "react";
+import { StyleSheet } from "react-native";
+import React from "react";
 import {
   Select,
   SelectTrigger,
