+++ common_components/Thumbnail.js
@@ -1,4 +1,4 @@
-import { Image, StyleSheet, Text, View } from "react-native";
+import { Image, StyleSheet } from "react-native";
 import React from "react";
 import utils from "@/core/utils";
