+++ screens/Auctions/components/PaymentMethod.js
@@ -1,4 +1,4 @@
-import { Image, StyleSheet, Text, TouchableOpacity, View } from "react-native";
+import { Image, StyleSheet, Text, TouchableOpacity } from "react-native";
 import React from "react";
 
 const PaymentMethod = ({ icon, label }) => {
