+++ screens/Chats/components/ChatInput.js
@@ -1,10 +1,4 @@
-import {
-  StyleSheet,
-  Text,
-  TextInput,
-  TouchableOpacity,
-  View,
-} from "react-native";
+import { StyleSheet, TextInput, TouchableOpacity, View } from "react-native";
 import React from "react";
 
 const ChatInput = ({ message, setMessage, onSend }) => {
